+++ server/database.go
@@ -444,6 +444,10 @@ func (db *Database) migrate() error {
 		return formatError(err, "")
 	}
 
+	if err := migrateSystemDuplicateDetection(db); err != nil {
+		return formatError(err, "")
+	}
+
 	return nil
 }